python/pyproject.toml
[tool.poetry]
name = "lospinguinos"
version = "0.1.0"
description = ""
authors = ["Vladislav <mribloko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"
flask = "^3.0.2"
flask-cors = "^4.0.0"
fastapi = "^0.109.2"
fastapi-sessions = "^0.3.2"
uuid = "^1.30"
requests = "^2.31.0"
uvicorn = "^0.27.1"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.2.0"
pre-commit = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
